add function in openweather.map
